Move page specific members to new class
